Reorganized the routes folder and now all routes are imported into app.js using require and the app variable is passed to each route folder.
