+++ app.js
@@ -1,17 +1,10 @@
 const express = require("express");
-const pool = require("./dbPool.js");
-const mysql = require('mysql2');
 const cookieParser = require('cookie-parser');
 const session = require('express-session');
 const passport = require('passport');
 const passportConfig = require('./config/passport')(passport);
 const flash = require('express-flash-messages');
 const app = express();
-
-const { handleLogout, postLogin, getLogin } = require('./routes/login');
-const { getFacebookLogin, handleFacebookLogin } = require('./routes/facebook_login');
-const { getGoogleLogin, handleGoogleLogin } = require('./routes/google_login');
-const { getRegister, submitRegister } = require('./routes/register');
 
 //Used to parse the body of a post request.
 app.use(express.urlencoded({ extended: true}));
@@ -29,217 +22,12 @@ app.use(passport.initialize());
 app.use(passport.session());
 app.use(flash());
 
-//middleware function
-const authCheck = (req, res, next) =>{
-    if(!req.user){
-        //if user is not logged in
-        res.redirect("/login");
-    }else{
-        //if logged in
-        next();
-    }
-};
-
-//login, registration & authentication routes
-app.post('/login', postLogin);
-app.get('/login', getLogin);
-app.get('/logout', handleLogout);
-
-app.get('/signup', getRegister);
-app.post('/signup', submitRegister);
-
-app.get('/auth/facebook', getFacebookLogin);
-app.get('/auth/facebook/callback', handleFacebookLogin);
-app.get('/auth/google', getGoogleLogin);
-app.get('/auth/google/callback', handleGoogleLogin);
-
-//Unprotected routes
-app.get("/login", function(req, res){
-    res.render("login");
-});
-
-app.get("/logout", function(req, res){
-    req.logOut()
-    res.redirect("/")
-});
-
-app.get("/", function(req, res){
-    res.render("index");
-});// "/"
-
-//protected routes
-app.get('/dashboard', authCheck, (req, res) => {
-    res.render("dashboard");
-});
-
-app.get('/sandbox', authCheck, (req, res) => {
-    res.render("sandbox");
-});
-
-app.get('/lecture', authCheck, (req, res) => {
-    res.render("lecture");
-});
-
-app.get("/chapter", authCheck, function(req, res){
-    res.render("chapter");
-});
-
-app.get("/admin", authCheck, function(req, res){
-    res.render("admin");
-}); // "/admin"
-
-app.get("/admin/chapters/add", authCheck, function(req, res){
-    res.render("addChapter");
-}); // "/admin/chapters/add"
-
-app.get("/admin/pages/add", authCheck, function(req, res){
-    res.render("addPage");
-}); // "/admin/pages/add"
-
-app.get("/admin/chapters/edit", authCheck, function(req, res){
-    res.render("editChapter");
-}); // "/admin/chapters/edit"
-
-app.get("/admin/pages/edit", authCheck, function(req, res){
-    res.render("editPage");
-}); // "/admin/pages/edit"
-
-app.get('/quizzes', authCheck, function(req , res){
-    res.render('quizzes');
-});
-
-
-app.get('/quizzes/:id', function(req , res){
-    res.render('quiz' + req.params.id);
-});
-
-/*
-Two APIs for getting chapters from database.
-1. /chapters?action=all
-    -Returns all data from the chapters table.
-2. /chapters?action=titles
-    -Returns the id and title for every chapter
-*/
-app.get("/chapters", function(req, res){
-    let sql;
-    let sqlParams = [];
-    switch (req.query.action) {
-        case "all":       sql = "SELECT * FROM chapters";
-                          break;
-        case "titles":    sql = "SELECT id, title FROM chapters";
-                          break;
-        case "single":    sql = "SELECT * from chapters WHERE id = ?";
-                          sqlParams = [req.query.id];
-                          break;
-        default:          res.status(400).send('Invalid API action');
-                          return;
-    }
-
-    pool.query(sql, sqlParams, function (err, rows, fields) {
-        if (err) {
-            console.log(err);
-            res.status(500).send('Internal database error');
-        } else {
-            // console.log(rows); //testing
-            res.send(rows);//sends data
-        }
-    });
-}); // "/chapters"
-
-//Receives post form request
-app.post("/admin/chapters/add", function(req, res){
-    let sql = "INSERT INTO chapters (title, description) VALUES (?,?)";
-    let sqlParams = [req.body.title, req.body.description];
-
-    pool.query(sql, sqlParams, function(err, rows, fields) {
-        if (err) console.log(err);
-        // console.log(rows); //testing
-        res.render("admin");
-    });
-}); // "/admin/chapters/add"
-
-//Receives post form request
-app.post("/admin/chapters/edit", function(req, res){
-    let sql = "UPDATE chapters SET title = ?, description = ? WHERE id = ?";
-    let sqlParams = [req.body.title, req.body.description, req.body.chapter];
-
-    pool.query(sql, sqlParams, function(err, rows, fields) {
-        if (err) console.log(err);
-        // console.log(rows); //testing
-        res.render("admin");
-    });
-}); // "/admin/chapters/edit"
-
-/*
-Three APIs for getting pages from database.
-1. /pages?action=page&id={}
-    -Returns a single page based upon the page_id
-2. /pages?action=chapter&id={}
-    -Returns an entire chapter of pages based upon the chapter_id
-3. /pages?action=titles
-    -Returns all titles in the format: page_title, page_id, chapter_title, chapter_id
-*/
-app.get("/pages", function(req, res){
-    let sql;
-    let sqlParams = [];
-
-    switch(req.query.action){
-        case "page":      sql = "SELECT * FROM pages WHERE id = ?";
-                          sqlParams.push(req.query.id);
-                          break;
-        case "chapter":   sql = "SELECT * FROM pages WHERE chapter_id = ?";
-                          sqlParams.push(req.query.id);
-                          break;
-        case "titles":    sql = `
-                          SELECT ch.id AS chapter_id,
-                                 ch.title AS chapter_title,
-                                 p.id AS page_id,
-                                 p.title AS page_title
-                          FROM pages AS p
-                          JOIN chapters AS ch
-                              ON ch.id = p.chapter_id
-                          ORDER BY ch.id, p.id
-                          `;
-                          break;
-        default:          res.status(400).send('Invalid API action');
-                          return;
-    }
-
-    pool.query(sql, sqlParams, function (err, rows, fields) {
-        if (err) {
-            console.log(err);
-            res.status(500).send('Internal database error');
-        } else {
-          // console.log(rows); //testing
-          res.send(rows);//sends
-        }
-    });
-}); // "/admin/pages"
-
-//Receives post form request
-app.post("/admin/pages/add", function(req, res){
-   let sql = "INSERT INTO pages (chapter_id, title, body) VALUES (?,?,?)";
-   let sqlParams = [req.body.chapter, req.body.title, req.body.body];
-
-   pool.query(sql, sqlParams, function(err, rows, fields) {
-        if (err) console.log(err);
-        // console.log(rows); //testing
-        res.render("admin");
-   });
-}); // "/admin/pages/add"
-
-//Receives post form request
-app.post("/admin/pages/edit", function(req, res){
-    let sql = "UPDATE pages SET title = ?, body = ? WHERE id = ?";
-    let sqlParams = [req.body.title, req.body.body, req.body.page];
-
-    pool.query(sql, sqlParams, function(err, rows, fields) {
-        if (err) console.log(err);
-        // console.log(rows); //testing
-        res.render("admin");
-    });
-}); // "/admin/pages/edit"
-
+//routes
+require("./routes/public.js")(app);
+require("./routes/login_register.js")(app);
+require("./routes/user.js")(app);
+require("./routes/admin.js")(app);
+require("./routes/api.js")(app);
 
 //starting server
 app.listen(process.env.PORT || 3000, process.env.IP, function(){
