+++ routes/public.js
@@ -1,0 +1,15 @@
+module.exports = (app) => {
+    //Unprotected routes
+    app.get("/login", function(req, res){
+        res.render("login");
+    });
+
+    app.get("/logout", function(req, res){
+        req.logOut()
+        res.redirect("/")
+    });
+
+    app.get("/", function(req, res){
+        res.render("index");
+    });// "/"
+};
